Replace int with specified types

+++ src/FPSManager.cpp
@@ -13,16 +13,16 @@ void FpsManager::syncFrameTime()
     }
 
     auto now = std::chrono::steady_clock::now();
-    currentFps = static_cast<int>(1.0 / std::chrono::duration<double>(now - timeAtStartOfFrame).count());
+    currentFps = static_cast<uint16_t>(1.0 / std::chrono::duration<double>(now - timeAtStartOfFrame).count());
     timeAtStartOfFrame = now;
 }
 
-void FpsManager::setTargetFps(int targetFps)
+void FpsManager::setTargetFps(uint16_t targetFps)
 {
     targetFrameTime = 1.0 / targetFps;
 }
 
-int FpsManager::getFps()
+uint16_t FpsManager::getFps()
 {
     return currentFps;
 }
